+++ aqt/installer.py
@@ -261,7 +261,15 @@ class Cli:
         if arch is not None and arch != "":
             return arch
         if os_name == "linux" and target == "desktop":
-            return "gcc_64"
+            try:
+                if Version(qt_version_or_spec) >= Version("6.7.0"):
+                    return "linux_gcc_64"
+                else:
+                    return "gcc_64"
+            except ValueError:
+                return "gcc_64"
+        elif os_name == "linux_arm64" and target == "desktop":
+            return "linux_gcc_arm64"
         elif os_name == "mac" and target == "desktop":
             return "clang_64"
         elif os_name == "mac" and target == "ios":
@@ -730,7 +738,7 @@ class Cli:
 
     def _set_install_tool_parser(self, install_tool_parser, *, is_legacy: bool):
         install_tool_parser.set_defaults(func=self.run_install_tool, is_legacy=is_legacy)
-        install_tool_parser.add_argument("host", choices=["linux", "mac", "windows"], help="host os name")
+        install_tool_parser.add_argument("host", choices=["linux", "linux_arm64", "mac", "windows"], help="host os name")
         if not is_legacy:
             install_tool_parser.add_argument(
                 "target",
@@ -787,7 +795,7 @@ class Cli:
 
         def make_parser_list_sde(cmd: str, desc: str, cmd_type: str):
             parser = subparsers.add_parser(cmd, description=desc)
-            parser.add_argument("host", choices=["linux", "mac", "windows"], help="host os name")
+            parser.add_argument("host", choices=["linux", "linux_arm64", "mac", "windows"], help="host os name")
             parser.add_argument(
                 "qt_version_spec",
                 metavar="(VERSION | SPECIFICATION)",
@@ -835,7 +843,7 @@ class Cli:
             "$ aqt list-qt mac desktop --archives 5.9.0 clang_64              # list archives in base Qt installation\n"
             "$ aqt list-qt mac desktop --archives 5.14.0 clang_64 debug_info  # list archives in debug_info module\n",
         )
-        list_parser.add_argument("host", choices=["linux", "mac", "windows"], help="host os name")
+        list_parser.add_argument("host", choices=["linux", "linux_arm64", "mac", "windows"], help="host os name")
         list_parser.add_argument(
             "target",
             nargs="?",
@@ -919,7 +927,7 @@ class Cli:
             "$ aqt list-tool mac desktop tools_ifw --long  # print tool variant names with metadata for QtIFW\n"
             "$ aqt list-tool mac desktop ifw --long        # print tool variant names with metadata for QtIFW\n",
         )
-        list_parser.add_argument("host", choices=["linux", "mac", "windows"], help="host os name")
+        list_parser.add_argument("host", choices=["linux", "linux_arm64", "mac", "windows"], help="host os name")
         list_parser.add_argument(
             "target",
             nargs="?",
@@ -1006,7 +1014,7 @@ class Cli:
         """
         if is_legacy:
             subparser.add_argument("qt_version", help='Qt version in the format of "5.X.Y"')
-        subparser.add_argument("host", choices=["linux", "mac", "windows"], help="host os name")
+        subparser.add_argument("host", choices=["linux", "linux_arm64", "mac", "windows"], help="host os name")
         if is_target_deprecated:
             subparser.add_argument(
                 "target",